feat: add role based actions

+++ src/main/java/ugr/dss/quick_shop/services/ProductsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ugr.dss.quick_shop.models.Product;
 
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
+public interface ProductsRepository extends JpaRepository<Product, Long> {
 
 }
